refactor

+++ src/modules/book/services/review.service.ts
@@ -4,7 +4,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { AddReviewDto } from '@/modules/book/dto/add-review.dto';
-import { DataSource } from 'typeorm';
+import { DataSource, EntityManager } from 'typeorm';
 import { Review } from '@/modules/book/entities/review.entity';
 import { User } from '@/modules/user/entities/user.entity';
 import { Book } from '@/modules/book/entities/book.entity';
@@ -12,6 +12,8 @@ import { UserErrors } from '@/modules/user/enums/errors.enum';
 import { BookErrors, ReviewErrors } from '@/modules/book/enums/errors.enum';
 import { UpdateReviewDto } from '@/modules/book/dto/update-review.dto';
 import { InjectDataSource } from '@nestjs/typeorm';
+import { AuthUser } from '@/modules/auth/types/auth-user.type';
+import { Roles } from '@/modules/user/enums/roles.enum';
 
 @Injectable()
 export class ReviewService {
@@ -29,27 +31,42 @@ export class ReviewService {
     );
   }
 
+  private async updateBookRating(manager: EntityManager, bookId: number) {
+    const bookRepository = manager.getRepository(Book);
+
+    const book = await bookRepository.findOne({
+      where: { id: bookId },
+      relations: { reviews: true },
+    });
+
+    if (!book) throw new NotFoundException(BookErrors.NOT_FOUND);
+
+    const newBookRating = this.calculateAverageRating(book.reviews);
+
+    const { affected } = await bookRepository.update(bookId, {
+      rating: newBookRating,
+    });
+
+    if (affected === 0) throw new BadRequestException(BookErrors.NOT_UPDATED);
+  }
+
   async addReview(userId: number, bookId: number, payload: AddReviewDto) {
     return await this.dataSource.transaction(async (manager) => {
       const userRepository = manager.getRepository(User);
       const bookRepository = manager.getRepository(Book);
       const reviewRepository = manager.getRepository(Review);
 
-      const existingReview = await reviewRepository.findOne({
+      const isExists = await reviewRepository.findOne({
         where: { user: { id: userId }, book: { id: bookId } },
       });
 
-      if (existingReview) {
-        throw new BadRequestException(ReviewErrors.ALREADY_EXISTS);
-      }
+      if (isExists) throw new BadRequestException(ReviewErrors.ALREADY_EXISTS);
 
       const user = await userRepository.findOneBy({ id: userId });
 
       if (!user) throw new NotFoundException(UserErrors.NOT_FOUND);
 
-      const book = await bookRepository.findOne({
-        where: { id: bookId },
-      });
+      const book = await bookRepository.findOneBy({ id: bookId });
 
       if (!book) throw new NotFoundException(BookErrors.NOT_FOUND);
 
@@ -64,20 +81,7 @@ export class ReviewService {
 
       const result = await reviewRepository.save(review);
 
-      const updatedBook = await bookRepository.findOne({
-        where: { id: bookId },
-        relations: { reviews: true },
-      });
-
-      if (!updatedBook) throw new NotFoundException(BookErrors.NOT_FOUND);
-
-      const bookRating = this.calculateAverageRating(updatedBook.reviews);
-
-      const { affected } = await bookRepository.update(bookId, {
-        rating: bookRating,
-      });
-
-      if (affected === 0) throw new BadRequestException(BookErrors.NOT_UPDATED);
+      await this.updateBookRating(manager, bookId);
 
       return result;
     });
@@ -92,19 +96,25 @@ export class ReviewService {
       const reviewRepository = manager.getRepository(Review);
 
       const review = await reviewRepository.findOne({
-        where: { id: reviewId, user: { id: userId } },
+        where: { id: reviewId },
+        relations: { book: true },
       });
 
       if (!review) throw new NotFoundException(ReviewErrors.NOT_FOUND);
+
+      if (userId !== review.user.id)
+        throw new BadRequestException(ReviewErrors.CANNOT_UPDATE);
 
       const { affected } = await reviewRepository.update(reviewId, payload);
 
       if (affected === 0)
         throw new BadRequestException(ReviewErrors.NOT_UPDATED);
 
-      const updated = await reviewRepository.findOne({
-        where: { id: reviewId, user: { id: userId } },
-      });
+      if (payload.rating !== review.rating) {
+        await this.updateBookRating(manager, review.book.id);
+      }
+
+      const updated = await reviewRepository.findOneBy({ id: reviewId });
 
       if (!updated) throw new NotFoundException(ReviewErrors.NOT_FOUND);
 
@@ -112,37 +122,23 @@ export class ReviewService {
     });
   }
 
-  async deleteReview(
-    userId: number,
-    bookId: number,
-    reviewId: number,
-  ): Promise<void> {
+  async deleteReview(user: AuthUser, reviewId: number): Promise<void> {
     return await this.dataSource.transaction(async (manager) => {
       const reviewRepository = manager.getRepository(Review);
-      const bookRepository = manager.getRepository(Book);
 
       const review = await reviewRepository.findOne({
-        where: { id: reviewId, user: { id: userId }, book: { id: bookId } },
+        where: { id: reviewId },
+        relations: { user: true, book: true },
       });
 
       if (!review) throw new NotFoundException(ReviewErrors.NOT_FOUND);
 
+      if (user.userId !== review.user.id && user.role !== Roles.ADMIN)
+        throw new BadRequestException(ReviewErrors.CANNOT_DELETE);
+
       await reviewRepository.remove(review);
 
-      const book = await bookRepository.findOne({
-        where: { id: bookId },
-        relations: { reviews: true },
-      });
-
-      if (!book) throw new NotFoundException(BookErrors.NOT_FOUND);
-
-      const bookRating = this.calculateAverageRating(book.reviews);
-
-      const { affected } = await bookRepository.update(bookId, {
-        rating: bookRating,
-      });
-
-      if (affected === 0) throw new BadRequestException(BookErrors.NOT_UPDATED);
+      await this.updateBookRating(manager, review.book.id);
     });
   }
 }
